Refactor socket middleware
