+++ src/store/middlewares/socket.js
@@ -1,5 +1,36 @@
+import io from '../../utils/socketShim';
 import actionTypes from '../../constants/actions';
-import { socketSetup, closeConnection } from '../../utils/api/socket';
+import { activePeerUpdate } from '../../actions/peers';
+import { SYNC_ACTIVE_INTERVAL, SYNC_INACTIVE_INTERVAL } from '../../constants/api';
+
+let connection;
+
+const socketSetup = (store) => {
+  let interval = SYNC_ACTIVE_INTERVAL;
+  const { ipc } = window;
+  if (ipc) {
+    ipc.on('blur', () => { interval = SYNC_INACTIVE_INTERVAL; });
+    ipc.on('focus', () => { interval = SYNC_ACTIVE_INTERVAL; });
+  }
+  connection = io.connect(`ws://${store.getState().peers.data.currentPeer}:${store.getState().peers.data.port}`);
+  connection.on('blocks/change', (block) => {
+    store.dispatch({
+      type: actionTypes.newBlockCreated,
+      data: { block, interval },
+    });
+  });
+  connection.on('disconnect', () => {
+    store.dispatch(activePeerUpdate({ online: false }));
+  });
+  connection.on('reconnect', () => {
+    store.dispatch(activePeerUpdate({ online: true }));
+  });
+};
+const closeConnection = () => {
+  if (connection) {
+    connection.close();
+  }
+};
 
 const socketMiddleware = store => (
   next => (action) => {
